Refactor all the remianing scenarios to be inline with the acceprtance criteria from the ticket
